Move GetContractVersionForVersion to contract package
