+++ test/e2e/clusterclass_rollout.go
@@ -1270,7 +1270,7 @@ func modifyControlPlaneViaClusterAndWait(ctx context.Context, input modifyContro
 		controlPlane, err := external.Get(ctx, mgmtClient, controlPlaneRef)
 		g.Expect(err).ToNot(HaveOccurred())
 
-		contractVersion, err := utilconversion.GetContractVersion(ctx, mgmtClient, controlPlane.GroupVersionKind())
+		contractVersion, err := contract.GetContractVersion(ctx, mgmtClient, controlPlane.GroupVersionKind())
 		g.Expect(err).ToNot(HaveOccurred())
 
 		// Verify that the fields from Cluster topology are set on the control plane.
